% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{assets}
\alias{assets}
\title{assets}
\usage{
assets(l)
}
\arguments{
\item{l}{the lines of an html file}
}
\value{
a vector of the css, js and png files called locally in the file
}
\description{
Assets
}
\author{
Thomas Rosendal
}
