% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{cwd_time}
\alias{cwd_time}
\title{time_series

This function creates a time series graph}
\usage{
cwd_time(path = system.file("sample_data_cwd.csv", package = "svamap"),
  encoding = "UTF-8", date_in = "Ankomstdatum", target = "Djurslagskod")
}
\arguments{
\item{path}{Path to file}

\item{encoding}{Ecoding of the text file}

\item{date_in}{dataset field with dates}

\item{target}{numeric value to be plotted}
}
\value{
A data.frame
}
\description{
Create a time series graph using dygraphs.js library
}
\author{
Giampaolo Cocca
}

