% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Table.R
\name{do_Table}
\alias{do_Table}
\title{do_Table}
\usage{
do_Table(x, disease = "myDisease", tabhead, targets = NULL,
  lengthpage = NULL, tocolor = NULL, colorPal = NULL, targetcol = NULL,
  initComplete = NULL, height = NULL, width = NULL, browse = FALSE,
  dir = tempdir())
}
\arguments{
\item{x}{an object of class data.frame}

\item{disease}{name string of the disease investigated}

\item{tabhead}{a character vector with the header of the table}

\item{targets}{vector of numbers defining position of the column to center. First column is 0}

\item{lengthpage}{number of records to display per page}

\item{tocolor}{the indices of the columns to be formatted (can be character, numeric, logical, or a formula of the form ~ V1 + V2, which is equivalent to c('V1', 'V2'))}

\item{colorPal}{a vector of color to use as cells' background}

\item{targetcol}{a character vector in the first column to be assigned (one-to-one) to 'colorPal'. The style will be applied to the entire row}

\item{initComplete}{JS snippet that modify background color and text color of the header}

\item{height}{the height of the table in pixel}

\item{width}{the width of the table in pixel}

\item{browse}{if TRUE it opens the map in a new page of the default browser}

\item{dir}{path where the map will be saved}
}
\value{
path to the table
}
\description{
Creates an .html table with custom style
}
\details{
The function is built on the top of R "DT" package.
Package bind "DataTables" a plug-in for the jQuery Javascript library.
}
\author{
Giampaolo Cocca
}

