% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{friends}
\alias{friends}
\title{friends}
\usage{
friends(df)
}
\arguments{
\item{df}{A dataframe read from the PPN data in rapportportalen}
}
\value{
list A list of dataframes equal in length to the total
    unique PPNs.
}
\description{
Find the friends of a PPN based on being reported in the same
ownership chain.
}
\details{
\code{friends} calculated the distance between PPNs in a network
based on ownership structure. A PPN has a reported owner or animal
manager. One PPN can have multiple people associated and these
people could also be associated with multiple PPNs. This function
finds the social network chains that exist in this data. For
example John and Alice own farm "A"; Alice and Jane own farm "B";
Jane and Bill own farm "C". Ann and Henry own farm "D". In this
case Farm "A" and "B" have a distance of 1 and A and C a distance
of 2. Farm D does not have any connections. These relationships
may reveal links between herds that are not reported in the animal
movement data.
}
\examples{
\dontrun{
df <- read.csv(file = "/media/ubuntu1/PPN_records.csv",
               sep=";", header=T, stringsAsFactors = FALSE,
               dec=",", encoding = "UTF-8")
df <- df[1:1000,]
friends <- friends(df)
}
}
\author{
Thomas Rosendal
}
