% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{generate_header}
\alias{generate_header}
\title{generate header}
\usage{
generate_header(beginning = c("<meta charset=\\"utf-8\\" />",
  "<meta http-equiv=\\"x-ua-compatible\\" content=\\"IE=edge\\" >",
  "<link rel=\\"stylesheet\\" href=\\"http://www.sva.se/assets/css/main.css\\" />"),
  paging = FALSE, ordering = FALSE, info = FALSE, searching = FALSE)
}
\arguments{
\item{beginning}{A free set of text lines. This is a character
vector that contains the text to be put in the html <head>
section of the page. The default calls the sva main.css and
makes the page compatible with IE and sets text encoding to
UTF-8}

\item{paging}{Either TRUE or FALSE. Do you want the table to have paging?}

\item{ordering}{Either TRUE or FALSE Do you want sorting?}

\item{info}{Either TRUE or FALSE Do you want info?}

\item{searching}{Either TRUE or FALSE Do you want a search field?}
}
\value{
A header object to be passed to the html_table function
}
\description{
A function that generates the header for the html table
}
\author{
Thomas Rosendal
}

