% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{html_table}
\alias{html_table}
\title{html_table}
\usage{
html_table(df,
  html.table.attributes = "id = \\"table1\\" class=\\"svatablegrayheader\\" style=\\"width: 100\%;\\" border=\\"0\\"",
  html_head = generate_header(), align = rep("left", ncol(df)),
  col.names = names(df), file = tempfile(), fragment = FALSE)
}
\arguments{
\item{df}{The dataframe to be written to html table form}

\item{html.table.attributes}{The attributes of the table. This is
a character vector that is placed in the <table> tag. Should
include the class(es) of the table. The default is an html
table with class 'svatablegrayheader' which is one of the
table classes defined in sva's main.css}

\item{html_head}{An object of class svamap.htmlheader. See generate_header() function}

\item{align}{The alignement of each column. This is a character
vector that indicates the alignment of each column in the
table. The default is all 'left'}

\item{col.names}{The names of the column heading in the
table. This is a character vector of the names to be used in
the resulting table. The default are the column names from df}

\item{file}{A path to a file. A valid path to write the html. The
default is a tempfile().}

\item{fragment}{Do you want just a fragment or a full html?}
}
\value{
A path to an html file
}
\description{
html_table
}
\details{
A function that takes a data.frame and writes to an html table
}
\author{
Thomas Rosendal
}

