% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{kommuner}
\alias{kommuner}
\title{Administrative (NUTS) units in Sweden (1 : 20 Million)}
\source{
\url{http://www.scb.se/Grupp/Produkter_Tjanster/Verktyg/_Dokument/Shape-svenska.zip}
}
\description{
The dataset contains administrative units (NUTS) in Sweden in
European terminology in scale 1 : 20 Million.
}
\details{
Reprojected to WGS84
}
\examples{

data(kommuner)
plot(kommuner)

}
\keyword{datasets}

