% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{kommuner_norge}
\alias{kommuner_norge}
\title{Administrative divisions 'kommuner' of Norway}
\source{
\url{http://data.kartverket.no/download/content/administrative-enheter-norge-wgs-84-hele-landet-geojson}
}
\description{
The dataset contains administrative units 'kommuner' in Norway
}
\keyword{datasets}
