% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{lan}
\alias{lan}
\title{Administrative units in Sweden (Lan)}
\source{
\url{http://www.scb.se/Grupp/Produkter_Tjanster/Verktyg/_Dokument/Shape-svenska.zip}
}
\description{
The dataset contains boundaries of lan in Sweden
}
\details{
Reprojected to WGS84
}
\keyword{datasets}
