% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_county.R
\name{match_to_county}
\alias{match_to_county}
\title{match_to_county}
\usage{
match_to_county(pts, polygons, id = "NUTS_ID")
}
\arguments{
\item{pts}{a spatial points dataframe}

\item{polygons}{and spatial polygons dataframe}

\item{id}{The id field in the polygondataset}
}
\value{
A spatial polygon dataframe
}
\description{
Function that takes data from point layer and applies it to the
appropriate polygons and returns a polygon layer with the chosen
summarized data
}
\author{
Thomas Rosendal
}

