% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{NUTS_20M}
\alias{NUTS_20M}
\title{Administrative (NUTS) units in Sweden (1 : 20 Million)}
\source{
\url{http://epp.eurostat.ec.europa.eu/portal/page/portal/gisco_Geographical_information_maps/popups/references/administrative_units_statistical_units_1}
}
\description{
The dataset contains administrative units (NUTS) in Sweden in
European terminology in scale 1 : 20 Million.
}
\details{
Reprojected to WGS84
}
\keyword{datasets}

