% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointleaf_map.R
\name{pointleaf_map}
\alias{pointleaf_map}
\title{pointleaf_map}
\usage{
pointleaf_map(mapdata, values, palette, title = NULL, labels, radius = 5,
  stroke = FALSE, popup = NULL,
  logo = "https://assets-cdn.github.com/images/modules/logos_page/GitHub-Logo.png",
  src = "remote", url = "https://github.com/", dir = tempdir(),
  disease = "myDisease", group = values, browse = FALSE)
}
\arguments{
\item{mapdata}{an object of class SpatialPointsDataFrame}

\item{values}{a vector of values of class "numeric" classifying the points}

\item{palette}{colors (HEX allowed) assigned to each of the corresponding classes (unique values of the argument 'values')}

\item{title}{the title of the legend}

\item{labels}{a vector of class "character" with labels to display in the legend}

\item{radius}{a numeric vector of radii for the circles (default to 5 pixels)}

\item{stroke}{whether to draw stroke along the borders of circles}

\item{popup}{text to be showed in the popup when clicking a point}

\item{logo}{logo to put in the topleft corner of the map}

\item{src}{character specifying the source location of the logo ("local" for images from the disk, "remote" for web image sources)}

\item{url}{the url to show when clicking the logo}

\item{dir}{path where the map will be saved}

\item{disease}{name of the html file. The suffix "_map.html" will be added to the provided name}

\item{group}{split the mapdata in different groups (default to unique values of the argument 'values')}

\item{browse}{if TRUE it opens the map in a new page of the default browser}
}
\value{
path to the map
}
\description{
Create a browsable points map
}
\details{
The function is built on the top of R "leaflet" and "mapview" packages.
Both packages bind the JavaScript library Leaflet.js
}
\author{
Giampaolo Cocca
}

