% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(path = system.file("sample_data_cwd.csv", package = "svamap"),
  encoding = "UTF-8",
  proj4str = "+proj=utm +zone=33 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
}
\arguments{
\item{path}{Path to file}

\item{encoding}{Ecoding of the text file}

\item{proj4str}{projection of points in file}
}
\value{
A data.frame
}
\description{
Read in the sample data and clean it
}
\author{
Thomas Rosendal
}

