% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_falken.R
\name{read_falken}
\alias{read_falken}
\title{read_falken}
\usage{
read_falken(path = system.file("sample_data_cwd.csv", package = "svamap"))
}
\arguments{
\item{path}{The path to the file with data}
}
\value{
A data.frame object
}
\description{
Read a standard Falken data query (.csv format)
}
\author{
SVA
}

