% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read_point_data}
\alias{read_point_data}
\title{read_point_data

This function reads in data from SVA's database on CWD
submittions. The default is to read in a fake dataset that has the
same strucuture.

FYI the following are some proj4string specs. RT90 is the default

RT90 is

"+init=epsg:3021
+proj=tmerc +lat_0=0
+lon_0=15.80827777777778
+k= ##' 1
+x_0=1500000
+y_0=0 +ellps=bessel
+towgs84=414.1,41.3,603.1,-0.855,2.141,-7.023,0
+units=m
+no_defs"

SWEREF99 is

"+proj=utm
+zone=33
+ellps=GRS80
+towgs84=0,0,0,0,0,0,0
+units=m
+no_defs"

WGS84 is

"+proj=longlat
+ellps=WGS84
+datum=WGS84
+no_defs"}
\usage{
read_point_data(path = system.file("sample_data_cwd.csv", package = "svamap"),
  encoding = "UTF-8", proj4str = paste0(c("+init=epsg:3021", "+proj=tmerc",
  "+lat_0=0", "+lon_0=15.80827777777778", "+k=1", "+x_0=1500000",
  "+y_0=0 +ellps=bessel", "+towgs84=414.1,41.3,603.1,-0.855,2.141,-7.023,0",
  "+units=m", "+no_defs"), collapse = " "),
  output_proj = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  long = "Gisy", lat = "Gisx")
}
\arguments{
\item{path}{Path to file}

\item{encoding}{Ecoding of the text file}

\item{proj4str}{projection of points in file}

\item{output_proj}{the projection of the returned dataframe (default WGS84)}

\item{long}{text string of the variable name that is longitude}

\item{lat}{text string of the variable name that is latitude}
}
\value{
A SpatialPointsDataframe
}
\description{
Read in point data and clean it
}
\author{
Thomas Rosendal
}
