% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{rough_kommune}
\alias{rough_kommune}
\title{Townships (kommune) that go into the water}
\description{
The dataset contains the boundaries of swedish kommuner extnding into the water
}
\keyword{datasets}
