% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_conversion.R
\name{rt90_sweref99}
\alias{rt90_sweref99}
\title{rt90_sweref99}
\usage{
rt90_sweref99(df)
}
\arguments{
\item{df}{a dataframe (read from the PPN dataset in RP)}
}
\value{
a dataframe
}
\description{
rt90_sweref99
}
\details{
Change all projected coordinated from sweref99 to RT90. The
coordinate systems do not overlap. All those points with a
Y-coordinate (the horizontal axis) of less than 1083427.290 are
sweref99 and should be reprojected to RT90. The other points are
already in RT90. The reason for the choice of RT90 is that the
polygon data that I have for Sweden is in RT90.

This function is designed to work on the PPN dataset from
rapportportalen (RP).
}
\author{
Thomas Rosendal
}
