% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_group.R
\name{samples_group}
\alias{samples_group}
\title{samples_group}
\usage{
samples_group(data_sample = read_falken(), breaks = c("week", "month",
  "day"), by = c("status", "material"))
}
\arguments{
\item{data_sample}{A data.frame object generated from the function
\code{read.gp()}}

\item{breaks}{Method used to group the date field (week, month,
day). Default to "week"}

\item{by}{Dataframe's field used to compute count and cumulative
sum. Default to "status"}
}
\value{
A data.frame object
}
\description{
Count and cumsum of data grouped by a specified time period
}
\details{
Reshape samples data extracted from Falken and cleaned through
\code{read.gp()} and gives back a data.frame object with count and
cumsum of data grouped by a specified time period
}
\author{
SVA
}

