% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{scaleLabel}
\alias{scaleLabel}
\title{scaleLabel}
\usage{
scaleLabel(display, labelString)
}
\arguments{
\item{display}{Should the label be visible}

\item{labelString}{The label}
}
\value{
A character vector
}
\description{
scaleLabel
}
\details{
generate the scaleLabel object
}
\author{
Thomas Rosendal
}
