% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{ticks}
\alias{ticks}
\title{ticks}
\usage{
ticks(min, max)
}
\arguments{
\item{min}{the minimum value}

\item{max}{the maximum value}
}
\value{
A character vector
}
\description{
ticks
}
\details{
Generate the tick object.
}
\author{
Thomas Rosendal
}
