% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{time_count}
\alias{time_count}
\title{time_count}
\usage{
time_count(time, breaks = c("days", "weeks", "months", "years"),
  count = c("freq", "cumul"), tmin = NULL, tmax = NULL)
}
\arguments{
\item{time}{a vector of dates}

\item{breaks}{one of c('days', 'weeks', 'months', 'years') to
split the data by}

\item{count}{'freq' for frequency count or 'cumul' for cumulative
count}

\item{tmin}{starting date (optional)}

\item{tmax}{ending date (optional)}
}
\value{
df dataframe containing the frequency or cumulative count of selected time breaks
}
\description{
time_count
}
\details{
Count the number of values in a date vector binned by various
intervals.
}
\author{
Thomas Rosendal
}
