% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{time_series_graph}
\alias{time_series_graph}
\title{time_series_graph

This function creates a time series graph using dygraphs.js
available in R through \code{library(dygraphs)}}
\usage{
time_series_graph(xts_obj, target, label, dir = tempdir(),
  disease = "my_disease", browse = FALSE)
}
\arguments{
\item{xts_obj}{Object of type xts. Possibly result from \code{svamap::write_time_series}}

\item{target}{numeric vector to plot}

\item{label}{character string to assign to y-axis and tooltip of the graph}

\item{dir}{path where the map will be saved}

\item{disease}{name of the html file. The suffix "_map.html" will be added to the provided name}

\item{browse}{if TRUE it opens the map in a new page of the default browser}
}
\value{
A data.frame
}
\description{
Create a time series graph using dygraphs.js library
}
\author{
Giampaolo Cocca
}

