% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_write.R
\name{time_series_write}
\alias{time_series_write}
\title{time_series_write

This function reads in data from SVA's database and produce 
an xts object with the cumsum of the confirmed cases}
\usage{
time_series_write(path = system.file("sample_data_cwd.csv", package =
  "svamap"), encoding = "UTF-8", date_in = "Ankomstdatum",
  target = "Status..numerisk.")
}
\arguments{
\item{path}{Path to file}

\item{encoding}{Ecoding of the text file}

\item{date_in}{dataset field with dates}

\item{target}{numeric value to be plotted}
}
\value{
A data.frame object
}
\description{
Generate a time series (xts) object
}
\author{
Giampaolo Cocca
}

