% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries_html}
\alias{timeseries_html}
\title{timeseries_html}
\usage{
timeseries_html(dataname = "data", datafilename = "data.js", yAxes)
}
\arguments{
\item{dataname}{The name of the data object to be loaded in the
html. This should match your call in the json function.}

\item{datafilename}{The name of the .js file that you need to load
from the html}

\item{yAxes}{The specification of the yaxis, of axes. This must be
of classy(Axes) == "yAxes" and is a list of objects generated
by the svamap::yAxis() function}
}
\value{
A character vector
}
\description{
timeseries_html
}
\details{
A method to generate the nessessary html for a timeseries plot
using chart.js.
}
\author{
Thomas Rosendal
}
