% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{Villreinomrader_norge}
\alias{Villreinomrader_norge}
\title{Divisions 'Villreinomrader' in Norway}
\source{
\url{http://trdfme01.miljodirektoratet.no/nedlasting/kartkatalog/Villreinomrader_norge.zip}
}
\description{
The dataset contains the boundaries of "Villreinomrader" in Norway
}
\keyword{datasets}
