% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{yAxes}
\alias{yAxes}
\title{yAxes}
\usage{
yAxes(x)
}
\arguments{
\item{x}{A list of yAxis objects}
}
\value{
a character vector of class 'yAxes'
}
\description{
yAxes
}
\details{
stick together you various yAxis objects into a yAxes object.
}
\author{
Thomas Rosendal
}
