% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Table.R
\name{do_Table}
\alias{do_Table}
\title{do_Table}
\usage{
do_Table(x, disease, header, tocolor, colorPal, target, browse, dir)
}
\arguments{
\item{x}{an object of class data.frame}

\item{disease}{name string of the disease investigated}

\item{header}{character vector with the header of the table}

\item{tocolor}{character vector with the name of the field(s) to format}

\item{colorPal}{color to use as cells background}

\item{target}{vector on which apply the style}

\item{browse}{if TRUE it opens the map in a new page of the default browser}

\item{dir}{path where the map will be saved}
}
\value{
path to the table
}
\description{
Create a an .html table with backgroundcolors
}
\details{
The function is built on the top of R "DT" package.
Package bind "DataTables" a plug-in for the jQuery Javascript library
}
\author{
Giampaolo Cocca
}

