% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svamap.R
\docType{data}
\name{kommuner}
\alias{kommuner}
\title{Administrative units in Sweden (Kommun)}
\source{
\url{http://www.scb.se/Grupp/Produkter_Tjanster/Verktyg/_Dokument/Shape-svenska.zip}
}
\description{
The dataset contains boundaries of kommun in Sweden
}
\details{
Reprojected to WGS84
}
\keyword{datasets}

