% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_map.R
\name{point_map}
\alias{point_map}
\title{point_map

FYI the following are some proj4string specs. RT90 is the default

RT90 is

"+init=epsg:3021 +proj=tmerc +lat_0=0 +lon_0=15.80827777777778 +k=1 +x_0=1500000 +y_0=0 +ellps=bessel +towgs84=414.1,41.3,603.1,-0.855,2.141,-7.023,0 +units=m +no_defs"

SWEREF99 is

"+proj=utm +zone=33 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"

WGS84 is

"+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"}
\usage{
point_map(pts,
  map_proj = "+init=epsg:3021 +proj=tmerc +lat_0=0 +lon_0=15.80827777777778 +k=1 +x_0=1500000 +y_0=0 +ellps=bessel +towgs84=414.1,41.3,603.1,-0.855,2.141,-7.023,0 +units=m +no_defs",
  basemap = "NUTS_20M", path_to_file = tempfile(fileext = ".pdf"),
  height = 7, width = 5, pch = 20, col = "#67841E", cex = 1)
}
\arguments{
\item{pts}{a spatialopointsdataframe}

\item{map_proj}{The projection of those output map (represented in proj4string)}

\item{basemap}{The name of the basemap}

\item{path_to_file}{The output file}

\item{height}{The height in inches}

\item{width}{The width in inches}

\item{pch}{The plotting character}

\item{col}{The col of the plotting character}

\item{cex}{The size of the plotting character}
}
\value{
A path to the pdf file of the map
}
\description{
Create a static map
}
\author{
Thomas Rosendal
}

