% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{time_series_graph}
\alias{time_series_graph}
\title{time_series_graph

This function creates a time series graph using dygraphs.js
available in R through \code{library(dygraphs)}}
\usage{
time_series_graph(xts_object = time_series_write()[[2]],
  target = colnames(time_series_write()[[2]][, 1]))
}
\arguments{
\item{xts_object}{foo}

\item{target}{numeric value to be plotted}
}
\value{
A data.frame
}
\description{
Create a time series graph using dygraphs.js library
}
\author{
Giampaolo Cocca
}

