% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{write_data}
\alias{write_data}
\title{write_data}
\usage{
write_data(object, varname = "data1", file = tempfile())
}
\arguments{
\item{object}{A spatial polygon or point dataframe}

\item{varname}{The name of the variable to be given to the geojson oject in javascript.}

\item{file}{The path to where the geojson will be written}
}
\value{
path to the file
}
\description{
Write a geojson file from a dataset
}
\author{
Thomas Rosendal
}

