% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{write_page}
\alias{write_page}
\title{write_page}
\usage{
write_page(data, path = tempdir(), overwrite = FALSE, browse = TRUE)
}
\arguments{
\item{data}{path to data}

\item{path}{path to new webpage}

\item{overwrite}{Do you want to overwrite files in the destination directory?}

\item{browse}{Pop the brwoser and view the page?}
}
\description{
write a webpage
}
\details{
Copy the files to a direcetory
}
\author{
Thomas Rosendal
}

