% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{write_page}
\alias{write_page}
\title{write_page}
\usage{
write_page(data, path = tempdir(), ftp = NULL, template = "CWD/map.html",
  owntemplate = NULL, overwrite = FALSE, browse = TRUE)
}
\arguments{
\item{data}{path to data}

\item{path}{path to new webpage does not end in a "/"}

\item{ftp}{NULL if you want to you a path to a directory,
otherwise specify the ftp server with credentials like:
"ftp://User:Password@servername/Destination". This is the
directory on the ftp server that the files will be placed. It
must end in a '/' and will be expanded when called by
ftpUpload from the RCurl library.}

\item{template}{the name of the map template. This is a .html file
in the inst directory of svamap}

\item{owntemplate}{path to you own template. This would include
everything your map needs but the data including the assets,
like .css, .js and .png files}

\item{overwrite}{Do you want to overwrite files in the destination
directory?}

\item{browse}{Pop the brwoser and view the page?}
}
\value{
The path to the map or NULL if ftp
}
\description{
write a webpage
}
\details{
Copy the files to a direcetory
}
\author{
Thomas Rosendal
}
